Infer workbook & user data seed to common
- Moved above said to be more generic
